+++ src/Stormancer.Plugins/Blobs/Stormancer.Server.Plugins.BlobStorage/Changelog.rst
@@ -14,6 +14,8 @@ Unreleased
 Added
 *****
 - Added get blob content API.
+- Added Staging & commit blocks API to upload big blobs.
+- Added IBlobStorageEventHandler extensibility point.
 
 0.1.0.6
 ----------